Add rubocop-rake
Use `task foo: [:bar]` rather than `task foo: :bar` because
`rubocop-rake` seems to treat an array of prerequisites as just
specifying prerequisites and not defining a new task. No idea why.

This was not a problem for `:default` as that doesn't need a
description. It was a problem for `:gem` because `rubocop-rake` can't
tell that `task gem: :spec` isn't defining a new task.